Make rpc.go so it can use a single link and send/receive multiple messages (#52)
Service Bus and Event Hubs both allow you to send multiple messages on the management link. We weren't using this and it meant that we were creating a separate link _per_ request.

This PR changes that so we pass in a unique message ID for each request, which allows us to demux the requests in the client, thus only using a single link.

This is how the other track 2 SDKs currently handle it.